GIAIC/20_DAYS_OF_COODING/Agent_using_OpenAI _Agents_SDK/pyproject.toml
[project]
name = "agent-using-openai-agents-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
]
